requirements.txt
numpy
pennylane
matplotlib
openqaoa-qiskit
docplex
cplex
pulp
